+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Minimum `PHP` is `7.1.3`
 - Moved handling request logic to dedicated class `PabloK\SupercacheBundle\Request\Handler`
 - Request/Response Handlers marked as `@internal`
+- Renamed `\PabloK\SupercacheBundle\Factory\ResponseFactory` to `\PabloK\SupercacheBundle\Response\ResponseFactory`
+  this can be a BC if you use class 
 
 ### Removed
 - `KernelListener` class - if you use this class then it is a BC break, but that class was meant for internal use only